requirements.txt
fastapi
uvicorn
loguru
lagom
python-multipart
python-jose[cryptography]
